backend/requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.7
psutil>=5.9
websockets>=12.0
apscheduler>=3.10
pygame>=2.5
sqlmodel>=0.0.16
python-multipart>=0.0.9
google-genai
python-dotenv
openai
ddgs
pytz
openwakeword
faster-whisper
piper-tts
pyaudio

